ARRISAPP-1462 Thunder: do not resubmit cleanup task
With _connectionCheckTimer set to 0 ('idletime' config setting)
when TriggerCleanup is called quickly, eg. because some
client disconnected & more than 1 channel has been closed,
it might happen that the 'cleanup task' (ChannelMap::_job) is
being added again before it had a chance to start execution,
and before it was removed from ThreadPool queue.
This results in the same object being added another time to
ThreadPool _queue, which results in ASSERT.
This fix makes sure that the task is not already in SUBMITTED
state before adding it again.

Change-Id: I202e4dc98a71e707a505c1525497274df4b5b800
Reviewed-on: https://gerrit.onemw.net/c/meta-thunder-rdkservices/+/127736
Tested-by: Jenkins <onemw-builder@onemw.net>
Tested-by: Tomasz Karczewski <tomasz.karczewski@redembedded.com>
Reviewed-by: Tomasz Karczewski <tomasz.karczewski@redembedded.com>
Reviewed-by: Abhayadev Sankaran <asankaran.ext@libertyglobal.com>